Create AbilityScoreWizard.xaml.cs
With the AbilityScoreWizard a tool was introduced that lets the user generate values for the six Ability Scores via virtual dice rolls instead of entering them manually. - Like in the original game a 6-sided die is rolled for each value/ Ability Score where the lowest result is ignored. - These values can then be re-arranged by the user in the wizard via 'Drag'n'Drop'.

The AbilityScoreWizard already uses the 'Race'-Class to add boni to the values/Ability Scores depending on the selected Character-Race and Subrace. Those boni are read with additional data about the Character-Race from a Data Base where all the Race-Options for Characters are stored.